Create README.md
WELCOME, and thanks for taking your time to check  out my repositories...

Here am building a custom google search engine, Primarily the request is from  the provided google url which is available on google cloud with the API KEY and API ID
With this API am you a using a free API key, you can make up to 100 queries per day for free. This translates to an average of approximately 4 queries per hour.